addons/images/count_metric/requirements.txt
APScheduler==3.10.0
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
coverage==7.1.0
exceptiongroup==1.1.0
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.5.0
google-cloud-core==2.3.2
google-cloud-pubsub==2.14.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
iniconfig==2.0.0
kafka-python==2.0.2
mock==5.0.1
packaging==23.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.2.1
pytest-cov==4.0.0
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rsa==4.9
six==1.16.0
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
google-cloud-datastore==2.15.1